+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 ## Change log
 ----------------------
 
+Version 3.5
+-------------
+
+ADDED:
+
+- added new test-dependency junit-jupiter (junit 5) in version 5.8.2
+
+CHANGED:
+
+- remove of test-dependency testng
+- update dependency of com.github.ben-manes.versions.gradle.plugin to new version 0.41.0
+- update of dependency jobj-core in new version 5.2
+
 Version 3.4
 -------------